breaking: remove `SSRManifest` from public types
