+++ .changeset/flat-cases-see.md
@@ -1,0 +1,6 @@
+---
+"@sveltejs/kit": major
+---
+
+breaking: remove `SSRManifest` from public types
+
